Estimate the MLE of a Mallows-Binomial distribution using the Greedy method

This function estimates the MLE of a Mallows-Binomial distribution using the Greedy method.

Greedy(rankings, ratings, M)

Arguments

rankings

A matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One ranking per row.

ratings

A matrix of ratings, one row per judge and one column per object.

M

Numeric specifying maximum (=worst quality) integer rating.

Value

A list with elements pi0, the estimated consensus ranking MLE, p, the estimated object quality parameter MLE, theta, the estimated scale parameter MLE, and numnodes, number of nodes traversed during algorithm and a measure of computational complexity. If multiple MLEs are found, pi0, p, and theta are returned a matrix elements, with one row per MLE.

Examples

data("ToyData1")
Greedy(ToyData1$rankings,ToyData1$ratings,ToyData1$M)
#> $pi0
#> [1] 1 2 3
#> 
#> $p
#> [1] 0.125 0.125 0.750
#> 
#> $theta
#> [1] 1e+08
#> 
#> $num_nodes
#> [1] 6
#>