Bootstrap Confidence Intervals for Mallows-Binomial parameters.

This function calculates confidence intervals for parameters in a Mallows-Binomial model using the nonparametric bootstrap.

ci_mb(
  rankings,
  ratings,
  M,
  interval = 0.9,
  nsamples = 50,
  all = FALSE,
  method = "ASTAR"
)

Arguments

rankings

A matrix of rankings, potentially with attribute "assignments" to signify separate reviewer assignments. One ranking per row.

ratings

A matrix of ratings, one row per judge and one column per object.

M

Numeric specifying maximum (=worst quality) integer rating.

interval

A numeric entry between 0 and 1 specifying the confidence interval (e.g., .90 indicates a 90% confidence interval). Defaults to 0.90.

nsamples

A numeric entry indicating desired number of bootstrap samples to be used when calculating confidence intervals. Defaults to 50.

all

A boolean indicating if estimated parameters from all bootstrap samples should be returned. Defaults to FALSE.

method

A character string indicating which estimation method to use when estimating parameters. Allowable options are currently "ASTAR", "Greedy", "GreedyLocal", and "FV". Defaults to exact search, "ASTAR".

Value

A list with elements ci, a matrix of confidence intervals for Mallows-Binomial parameters, ci_ranks, a matrix of confidence intervals for object ranks, bootstrap_pi0, a matrix of bootstrap consensus rankings (returned only if all==TRUE), and bootstrap_ptheta, a matrix of bootstrap estimates of (p,theta) (returned only if all==TRUE).

Examples

data("ToyData1")
ci_mb(ToyData1$rankings,ToyData1$ratings,ToyData1$M,method="ASTAR",all=TRUE)
#> $ci
#>            p1        p2   p3 theta
#> 5%  0.0781250 0.0781250 0.75 1e+08
#> 95% 0.1804687 0.1804687 0.75 1e+08
#> 
#> $ci_ranks
#>     Object1 Object2 Object3
#> 5%        1       2       3
#> 95%       1       2       3
#> 
#> $bootstrap_pi0
#>       [,1] [,2] [,3]
#>  [1,]    1    2    3
#>  [2,]    1    2    3
#>  [3,]    1    2    3
#>  [4,]    1    2    3
#>  [5,]    1    2    3
#>  [6,]    1    2    3
#>  [7,]    1    2    3
#>  [8,]    1    2    3
#>  [9,]    1    2    3
#> [10,]    1    2    3
#> [11,]    1    2    3
#> [12,]    1    2    3
#> [13,]    1    2    3
#> [14,]    1    2    3
#> [15,]    1    2    3
#> [16,]    1    2    3
#> [17,]    1    2    3
#> [18,]    1    2    3
#> [19,]    1    2    3
#> [20,]    1    2    3
#> [21,]    1    2    3
#> [22,]    1    2    3
#> [23,]    1    2    3
#> [24,]    1    2    3
#> [25,]    1    2    3
#> [26,]    1    2    3
#> [27,]    1    2    3
#> [28,]    1    2    3
#> [29,]    1    2    3
#> [30,]    1    2    3
#> [31,]    1    2    3
#> [32,]    1    2    3
#> [33,]    1    2    3
#> [34,]    1    2    3
#> [35,]    1    2    3
#> [36,]    1    2    3
#> [37,]    1    2    3
#> [38,]    1    2    3
#> [39,]    1    2    3
#> [40,]    1    2    3
#> [41,]    1    2    3
#> [42,]    1    2    3
#> [43,]    1    2    3
#> [44,]    1    2    3
#> [45,]    1    2    3
#> [46,]    1    2    3
#> [47,]    1    2    3
#> [48,]    1    2    3
#> [49,]    1    2    3
#> [50,]    1    2    3
#> 
#> $bootstrap_ptheta
#>             p1       p2   p3 theta
#>  [1,] 0.140625 0.140625 0.75 1e+08
#>  [2,] 0.187500 0.187500 0.75 1e+08
#>  [3,] 0.140625 0.140625 0.75 1e+08
#>  [4,] 0.125000 0.125000 0.75 1e+08
#>  [5,] 0.109375 0.109375 0.75 1e+08
#>  [6,] 0.187500 0.187500 0.75 1e+08
#>  [7,] 0.109375 0.109375 0.75 1e+08
#>  [8,] 0.171875 0.171875 0.75 1e+08
#>  [9,] 0.140625 0.140625 0.75 1e+08
#> [10,] 0.062500 0.062500 0.75 1e+08
#> [11,] 0.140625 0.140625 0.75 1e+08
#> [12,] 0.109375 0.109375 0.75 1e+08
#> [13,] 0.140625 0.140625 0.75 1e+08
#> [14,] 0.156250 0.156250 0.75 1e+08
#> [15,] 0.062500 0.062500 0.75 1e+08
#> [16,] 0.171875 0.171875 0.75 1e+08
#> [17,] 0.140625 0.140625 0.75 1e+08
#> [18,] 0.171875 0.171875 0.75 1e+08
#> [19,] 0.078125 0.078125 0.75 1e+08
#> [20,] 0.078125 0.078125 0.75 1e+08
#> [21,] 0.156250 0.156250 0.75 1e+08
#> [22,] 0.109375 0.109375 0.75 1e+08
#> [23,] 0.125000 0.125000 0.75 1e+08
#> [24,] 0.203125 0.203125 0.75 1e+08
#> [25,] 0.140625 0.140625 0.75 1e+08
#> [26,] 0.125000 0.125000 0.75 1e+08
#> [27,] 0.125000 0.125000 0.75 1e+08
#> [28,] 0.125000 0.125000 0.75 1e+08
#> [29,] 0.125000 0.125000 0.75 1e+08
#> [30,] 0.125000 0.125000 0.75 1e+08
#> [31,] 0.109375 0.109375 0.75 1e+08
#> [32,] 0.125000 0.125000 0.75 1e+08
#> [33,] 0.093750 0.093750 0.75 1e+08
#> [34,] 0.125000 0.125000 0.75 1e+08
#> [35,] 0.109375 0.109375 0.75 1e+08
#> [36,] 0.156250 0.156250 0.75 1e+08
#> [37,] 0.140625 0.140625 0.75 1e+08
#> [38,] 0.125000 0.125000 0.75 1e+08
#> [39,] 0.140625 0.140625 0.75 1e+08
#> [40,] 0.093750 0.093750 0.75 1e+08
#> [41,] 0.171875 0.171875 0.75 1e+08
#> [42,] 0.125000 0.125000 0.75 1e+08
#> [43,] 0.078125 0.078125 0.75 1e+08
#> [44,] 0.171875 0.171875 0.75 1e+08
#> [45,] 0.125000 0.125000 0.75 1e+08
#> [46,] 0.093750 0.093750 0.75 1e+08
#> [47,] 0.140625 0.140625 0.75 1e+08
#> [48,] 0.140625 0.140625 0.75 1e+08
#> [49,] 0.171875 0.171875 0.75 1e+08
#> [50,] 0.156250 0.156250 0.75 1e+08
#>